Add experimental support for the colorlight 5a-75b board
Changed the top.v in a migen/litex based wrapper. This should make
it quire easy to use any of the boards in litex-boards repository.

Experimental in sofar that bit stream is generated for Lattice ecp5
but no serial activity detected.
Stil have to figure out how to get some simulation traces.

Signed-off-by: Henk Vergonet <henk.vergonet@gmail.com>